reorganize code (#338)
