+++ packages/brain/src/modules/cron/index.ts
@@ -1,4 +1,4 @@
 export { CronJob } from "./cron.js";
-export { reloadValidators } from "./reloadValidators.js";
-export { sendProofsOfValidation } from "./sendProofsOfValidation.js";
-export { trackValidatorsPerformance, getSecondsToNextEpoch } from "./trackValidatorsPerformance.js";
+export { reloadValidators } from "./reloadValidators/index.js";
+export { sendProofsOfValidation } from "./sendProofsOfValidation/index.js";
+export { trackValidatorsPerformance, getSecondsToNextEpoch } from "./trackValidatorsPerformance/index.js";
