+++ packages/brain/src/modules/cron/trackValidatorsPerformance/logPrefix.ts
@@ -1,0 +1,1 @@
+export const logPrefix = "[CRON - trackValidatorsPerformance]: ";
